refactor: packages to views
