+++ ControleAssociacao/src/main/java/com/mycompany/controleassociacao/views/core/Login.java
@@ -2,7 +2,7 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JFrame.java to edit this template
  */
-package com.mycompany.controleassociacao.views;
+package com.mycompany.controleassociacao.views.core;
 
 import com.mycompany.controleassociacao.controllers.LoginController;
 import javax.swing.JOptionPane;
