+++ ControleAssociacao/src/main/java/com/mycompany/controleassociacao/views/member/CreateMember.java
@@ -2,7 +2,7 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JPanel.java to edit this template
  */
-package com.mycompany.controleassociacao.views;
+package com.mycompany.controleassociacao.views.member;
 
 import com.mycompany.controleassociacao.controllers.MemberController;
 import com.mycompany.controleassociacao.utils.Validate;
